requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.8.0
groq>=0.11.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0
langchain-community>=0.0.20

# Vector Stores
chromadb>=0.4.22
pinecone-client>=3.0.0
weaviate-client>=4.4.0

# Embeddings & ML
sentence-transformers>=2.3.0
tiktoken>=0.5.0
cohere>=4.47
numpy>=1.26.0

# HTTP & Async
httpx>=0.26.0
aiofiles>=23.2.0
tenacity>=8.2.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Logging
structlog>=24.1.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
